[feat] - réorganisation et initialisation curseur
