+++ Assets/Scripts/HealthBar.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5f250e3b3ffcaaf4084462d6a9edbb86
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
