+++ Assets/Scripts/HealthBar/HealthBar.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 83baabfc20433f24b89d684a79036862
+guid: dcd55757ab643d14ea1ecf18b6f3eb0a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
