+++ Assets/Scripts/HealthBar/HealthBarColor.cs
@@ -1,0 +1,3 @@
+
+using UnityEngine;
+public enum HealthBarColor { RED, GREEN, BLUE };
