+++ Assets/Scripts/HealthBar/HealthBarColor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 09b892b0b35ff8043ab41f7395455370
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
