+++ Assets/Scripts/HealthBar/HealthBarResponse.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e737ed3469e71c84a951a71fb5598d66
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
